+++ app/src/main/java/im/ene/lab/toro/sample/fragment/DeadlySimpleListFragment.java
@@ -36,7 +36,7 @@ import im.ene.lab.toro.ToroViewHolder;
 import im.ene.lab.toro.player.MediaSource;
 import im.ene.lab.toro.player.PlaybackException;
 import im.ene.lab.toro.player.TrMediaPlayer;
-import im.ene.lab.toro.player.widget.MediaPlayerView;
+import im.ene.lab.toro.player.widget.VideoPlayerView;
 import im.ene.lab.toro.sample.R;
 import im.ene.lab.toro.sample.data.SimpleVideoObject;
 import im.ene.lab.toro.sample.data.VideoSource;
@@ -92,7 +92,7 @@ public class DeadlySimpleListFragment extends Fragment {
 
     private static final int LAYOUT_RES = R.layout.vh_toro_video_deadly_simple;
 
-    private MediaPlayerView mVideoView;
+    private VideoPlayerView mVideoView;
     private ImageView mThumbnail;
 
     private SimpleVideoObject mItem;
@@ -100,7 +100,7 @@ public class DeadlySimpleListFragment extends Fragment {
 
     public SimpleViewHolder(View itemView) {
       super(itemView);
-      mVideoView = (MediaPlayerView) itemView.findViewById(R.id.video);
+      mVideoView = (VideoPlayerView) itemView.findViewById(R.id.video);
       mThumbnail = (ImageView) itemView.findViewById(R.id.thumbnail);
       mVideoView.setOnPreparedListener(this);
       mVideoView.setOnCompletionListener(this);
