+++ app/src/main/java/im/ene/lab/toro/sample/viewholder/HorizontalMixedToroVideoViewHolder.java
@@ -27,7 +27,7 @@ import com.squareup.picasso.Picasso;
 import im.ene.lab.toro.ToroVideoViewHolder;
 import im.ene.lab.toro.player.PlaybackException;
 import im.ene.lab.toro.player.TrMediaPlayer;
-import im.ene.lab.toro.player.widget.MediaPlayerView;
+import im.ene.lab.toro.player.widget.VideoPlayerView;
 import im.ene.lab.toro.sample.R;
 import im.ene.lab.toro.sample.data.SimpleVideoObject;
 import im.ene.lab.toro.sample.util.Util;
@@ -50,8 +50,8 @@ public class HorizontalMixedToroVideoViewHolder extends ToroVideoViewHolder {
     mInfo = (TextView) itemView.findViewById(R.id.info);
   }
 
-  @Override protected MediaPlayerView findVideoView(View itemView) {
-    return (MediaPlayerView) itemView.findViewById(R.id.video);
+  @Override protected VideoPlayerView findVideoView(View itemView) {
+    return (VideoPlayerView) itemView.findViewById(R.id.video);
   }
 
   private SimpleVideoObject mItem;
