+++ app/src/main/java/im/ene/lab/toro/sample/viewholder/SimpleToroVideoViewHolder.java
@@ -27,8 +27,8 @@ import com.squareup.picasso.Picasso;
 import im.ene.lab.toro.ToroVideoViewHolder;
 import im.ene.lab.toro.player.PlaybackException;
 import im.ene.lab.toro.player.TrMediaPlayer;
-import im.ene.lab.toro.player.widget.MediaPlayerView;
 import im.ene.lab.toro.player.widget.VideoPlayerView;
+import im.ene.lab.toro.player.widget.TextureVideoView;
 import im.ene.lab.toro.sample.R;
 import im.ene.lab.toro.sample.data.SimpleVideoObject;
 import im.ene.lab.toro.sample.facebook.TrackablePlayer;
@@ -38,7 +38,7 @@ import im.ene.lab.toro.sample.util.Util;
  * Created by eneim on 1/30/16.
  */
 public class SimpleToroVideoViewHolder extends ToroVideoViewHolder
-    implements TrackablePlayer, VideoPlayerView.OnReleasedListener {
+    implements TrackablePlayer, TextureVideoView.OnReleasedListener {
 
   private final String TAG = getClass().getSimpleName();
 
@@ -54,8 +54,8 @@ public class SimpleToroVideoViewHolder extends ToroVideoViewHolder
     // mVideoView.setOnReleasedListener(this);
   }
 
-  @Override protected MediaPlayerView findVideoView(View itemView) {
-    return (MediaPlayerView) itemView.findViewById(R.id.video);
+  @Override protected VideoPlayerView findVideoView(View itemView) {
+    return (VideoPlayerView) itemView.findViewById(R.id.video);
   }
 
   @Override public void setOnItemClickListener(View.OnClickListener listener) {
