+++ app/src/main/res/layout/sample_videoplayer.xml
@@ -21,7 +21,7 @@
     android:orientation="vertical"
     >
 
-  <im.ene.lab.toro.player.widget.MediaPlayerView
+  <im.ene.lab.toro.player.widget.VideoPlayerView
       android:id="@+id/player_view"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
