+++ app/src/main/res/layout/vh_toro_video_complicated.xml
@@ -73,7 +73,7 @@
         android:background="#20000000"
         >
 
-      <im.ene.lab.toro.player.widget.MediaPlayerView
+      <im.ene.lab.toro.player.widget.VideoPlayerView
           android:id="@+id/video"
           android:layout_width="match_parent"
           android:layout_height="180dp"
