+++ app/src/main/res/layout/vh_toro_video_complicated_horizontal.xml
@@ -76,7 +76,7 @@
         android:theme="@style/Theme.AppCompat"
         >
 
-      <im.ene.lab.toro.player.widget.MediaPlayerView
+      <im.ene.lab.toro.player.widget.VideoPlayerView
           android:id="@+id/video"
           android:layout_width="match_parent"
           android:layout_height="match_parent"
