+++ app/src/main/res/layout/vh_toro_video_simple.xml
@@ -24,7 +24,7 @@
     android:padding="4dp"
     >
 
-  <im.ene.lab.toro.player.widget.MediaPlayerView
+  <im.ene.lab.toro.player.widget.VideoPlayerView
       android:id="@+id/video"
       android:layout_width="match_parent"
       android:layout_height="180dp"
