+++ toro-player/src/main/java/im/ene/lab/toro/player/widget/VideoPlayerView.java
@@ -16,247 +16,402 @@
 
 package im.ene.lab.toro.player.widget;
 
+import android.Manifest;
 import android.annotation.TargetApi;
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.res.Resources;
+import android.content.pm.PackageManager;
 import android.graphics.SurfaceTexture;
-import android.media.AudioManager;
-import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Build;
 import android.support.annotation.FloatRange;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import android.support.v4.app.ActivityCompat;
 import android.util.AttributeSet;
-import android.util.Log;
-import android.view.KeyEvent;
-import android.view.MotionEvent;
 import android.view.Surface;
 import android.view.TextureView;
-import android.view.View;
-import android.view.accessibility.AccessibilityEvent;
-import android.view.accessibility.AccessibilityNodeInfo;
+import com.google.android.exoplayer.ExoPlayer;
+import com.google.android.exoplayer.audio.AudioCapabilities;
+import com.google.android.exoplayer.audio.AudioCapabilitiesReceiver;
+import com.google.android.exoplayer.metadata.id3.Id3Frame;
+import com.google.android.exoplayer.text.Cue;
+import com.google.android.exoplayer.util.Util;
+import im.ene.lab.toro.player.BuildConfig;
 import im.ene.lab.toro.player.MediaSource;
 import im.ene.lab.toro.player.PlaybackException;
 import im.ene.lab.toro.player.PlaybackInfo;
 import im.ene.lab.toro.player.TrMediaPlayer;
+import im.ene.lab.toro.player.internal.EventLogger;
+import im.ene.lab.toro.player.internal.ExoMediaPlayer;
+import im.ene.lab.toro.player.internal.RendererBuilderFactory;
 import im.ene.lab.toro.player.listener.OnBufferingUpdateListener;
 import im.ene.lab.toro.player.listener.OnCompletionListener;
 import im.ene.lab.toro.player.listener.OnErrorListener;
 import im.ene.lab.toro.player.listener.OnInfoListener;
 import im.ene.lab.toro.player.listener.OnPreparedListener;
-import im.ene.lab.toro.player.listener.OnVideoSizeChangedListener;
-import java.io.IOException;
-import java.util.Map;
+import java.util.List;
 
 /**
- * Displays a video file.  The TextureVideoView class
- * can load images from various sources (such as resources or content
- * providers), takes care of computing its measurement from the video so that
- * it can be used in any layout manager, and provides various display options
- * such as scaling and tinting.<p>
- *
- * <em>Note: VideoView does not retain its full state when going into the
- * background.</em>  In particular, it does not restore the current play state,
- * play position or selected tracks.  Applications should
- * save and restore these on their own in
- * {@link android.app.Activity#onSaveInstanceState} and
- * {@link android.app.Activity#onRestoreInstanceState}.<p>
- * Also note that the audio session id (from {@link #getAudioSessionId}) may
- * change from its previously returned value when the VideoView is restored.<p>
- *
- * This code is based on the official Android sources for 6.0.1_r10 with the following differences:
- * <ol>
- * <li>extends {@link android.view.TextureView} instead of a {@link android.view.SurfaceView}
- * allowing proper view animations</li>
- * <li>removes code that uses hidden APIs and thus is not available (e.g. subtitle support)</li>
- * </ol>
+ * Created by eneim on 6/4/16.
  */
-@Deprecated
 public class VideoPlayerView extends TextureView implements TrMediaPlayer.IMediaPlayer {
 
-  public interface OnReleasedListener {
+  private static final String TAG = "MediaPlayerView";
 
-    /**
-     * Called right before {@link #release(boolean)} )} get called with true
-     * parameter
-     *
-     * @param video current Video Uri
-     * @param position latest playback position right before releasing
-     * @param duration latest playback video's duration right before releasing
-     */
-    void onReleased(@Nullable Uri video, long position, long duration);
+  /**
+   * The surface view will not resize itself if the fractional difference
+   * between its default aspect ratio and the aspect ratio of the video falls
+   * below this threshold.
+   */
+  private static final float MAX_ASPECT_RATIO_DEFORMATION_PERCENT = 0.01f;
+
+  private AudioCapabilitiesReceiver.Listener audioCapabilitiesListener =
+      new AudioCapabilitiesReceiver.Listener() {
+        @Override public void onAudioCapabilitiesChanged(AudioCapabilities audioCapabilities) {
+          maybeNotifyAudioCapabilitiesChanged(audioCapabilities);
+        }
+      };
+
+  private SurfaceTextureListener mSurfaceTextureListener = new SurfaceTextureListener() {
+    @Override public void onSurfaceTextureAvailable(SurfaceTexture surface, int width, int height) {
+      VideoPlayerView.this.mSurface = new Surface(surface);
+      if (mMediaPlayer != null) {
+        mMediaPlayer.setSurface(VideoPlayerView.this.mSurface);
+        if (!mPlayerNeedsPrepare) {
+          mMediaPlayer.seekTo(mPlayerPosition);
+          mMediaPlayer.setPlayWhenReady(mPlayRequested);
+        }
+      } else {
+        preparePlayer(mPlayRequested);
+      }
+    }
+
+    @Override
+    public void onSurfaceTextureSizeChanged(SurfaceTexture surface, int width, int height) {
+
+    }
+
+    @Override public boolean onSurfaceTextureDestroyed(SurfaceTexture surface) {
+      if (!mBackgroundAudioEnabled) {
+        releasePlayer();
+      } else {
+        if (mMediaPlayer != null) {
+          mMediaPlayer.setBackgrounded(true);
+        }
+      }
+
+      if (mMediaPlayer != null) {
+        mMediaPlayer.blockingClearSurface();
+      }
+      VideoPlayerView.this.mSurface = null;
+      return true;
+    }
+
+    @Override public void onSurfaceTextureUpdated(SurfaceTexture surface) {
+
+    }
+  };
+
+  private final ExoMediaPlayer.Listener mPlayerListener = new ExoMediaPlayer.Listener() {
+    @Override public void onStateChanged(boolean playWhenReady, int playbackState) {
+      mPlaybackState = playbackState;
+      setKeepScreenOn(isInPlayableState());
+    }
+
+    @Override public void onError(Exception e) {
+      // Do nothing here
+    }
+
+    @Override public void onVideoSizeChanged(int width, int height, int unAppliedRotationDegrees,
+        float pixelWidthHeightRatio) {
+      if (getSurfaceTexture() != null) {
+        setVideoWidthHeightRatio(height == 0 ? 1 : (width * pixelWidthHeightRatio) / height);
+      }
+    }
+  };
+
+  private boolean isInPlayableState() {
+    return !mPlayerNeedsPrepare && (mPlaybackState != ExoMediaPlayer.STATE_IDLE) && (mPlaybackState
+        != ExoMediaPlayer.STATE_PREPARING) && (mPlaybackState != ExoMediaPlayer.STATE_ENDED);
   }
 
-  private String TAG = "TextureVideoView";
-  // settable by the client
   private Uri mUri;
-  private Map<String, String> mHeaders;
+  private AudioCapabilitiesReceiver mAudioCapabilitiesReceiver;
+  private AudioCapabilities mAudioCapabilities;
+  private ExoMediaPlayer mMediaPlayer;
+  private Surface mSurface;
 
-  // all possible internal states
-  private static final int STATE_ERROR = -1;
-  private static final int STATE_IDLE = 0;
-  private static final int STATE_PREPARING = 1;
-  private static final int STATE_PREPARED = 2;
-  private static final int STATE_PLAYING = 3;
-  private static final int STATE_PAUSED = 4;
-  private static final int STATE_PLAYBACK_COMPLETED = 5;
+  /**
+   * The ratio of the width and height of the video.
+   */
+  private float mVideoWidthHeightAspectRatio;
+  private long mPlayerPosition;
 
-  // mCurrentState is a TextureVideoView object's current state.
-  // mTargetState is the state that a method caller intends to reach.
-  // For instance, regardless the TextureVideoView object's current state,
-  // calling pause() intends to bring the object to a target state
-  // of STATE_PAUSED.
-  private int mCurrentState = STATE_IDLE;
-  private int mTargetState = STATE_IDLE;
-
-  // All the stuff we need for playing and showing a video
-  private Surface mSurface = null;
-  private TrMediaPlayer mMediaPlayer = null;
-  private int mAudioSession;
-  private int mVideoWidth;
-  private int mVideoHeight;
-
+  private int mPlaybackState;
+  private boolean mPlayerNeedsPrepare;
+  private boolean mPlayRequested = false;
   private boolean mBackgroundAudioEnabled = false;
 
-  private TrMediaPlayer.Controller mController;
-
-  private OnCompletionListener mOnCompletionListener;
   private OnPreparedListener mOnPreparedListener;
+  private OnCompletionListener mOnCompletionListener;
   private OnErrorListener mOnErrorListener;
   private OnInfoListener mOnInfoListener;
-  private OnReleasedListener mOnReleaseListener;
+  private OnBufferingUpdateListener mOnBufferingUpdateListener;
 
-  private int mCurrentBufferPercentage;
-  private long mSeekWhenPrepared;  // recording the seek position while preparing
+  // DEBUG
+  private EventLogger mEventLogger;
+
+  private ExoMediaPlayer.CaptionListener mCaptionListener;
+  private ExoMediaPlayer.Id3MetadataListener mId3MetadataListener;
+
+  private ExoMediaPlayerHelper mExoMediaPlayerHelper = new ExoMediaPlayerHelper() {
+    @Override public void onCues(List<Cue> cues) {
+      if (mCaptionListener != null) {
+        mCaptionListener.onCues(cues);
+      }
+    }
+
+    @Override public void onId3Metadata(List<Id3Frame> id3Frames) {
+      if (mId3MetadataListener != null) {
+        mId3MetadataListener.onId3Metadata(id3Frames);
+      }
+    }
+  };
+
+  public void setOnPreparedListener(OnPreparedListener onPreparedListener) {
+    this.mOnPreparedListener = onPreparedListener;
+  }
+
+  public void setOnCompletionListener(OnCompletionListener onCompletionListener) {
+    this.mOnCompletionListener = onCompletionListener;
+  }
+
+  public void setOnErrorListener(OnErrorListener onErrorListener) {
+    this.mOnErrorListener = onErrorListener;
+  }
+
+  public void setOnInfoListener(OnInfoListener onInfoListener) {
+    this.mOnInfoListener = onInfoListener;
+  }
+
+  public void setOnBufferingUpdateListener(OnBufferingUpdateListener onBufferingUpdateListener) {
+    this.mOnBufferingUpdateListener = onBufferingUpdateListener;
+  }
+
+  public void setCaptionListener(ExoMediaPlayer.CaptionListener listener) {
+    this.mCaptionListener = listener;
+  }
+
+  public void setId3MetadataListener(ExoMediaPlayer.Id3MetadataListener listener) {
+    this.mId3MetadataListener = listener;
+  }
+
+  private OnPreparedListener onPreparedListenerDelegate = new OnPreparedListener() {
+    @Override public void onPrepared(TrMediaPlayer mp) {
+      if (mOnPreparedListener != null) {
+        mOnPreparedListener.onPrepared(mp);
+      }
+    }
+  };
+
+  private OnCompletionListener onCompletionListenerDelegate = new OnCompletionListener() {
+    @Override public void onCompletion(TrMediaPlayer mp) {
+      mPlayRequested = false;
+      releasePlayer();
+      mPlayerPosition = 0;
+      if (mOnCompletionListener != null) {
+        mOnCompletionListener.onCompletion(mp);
+      }
+    }
+  };
+
+  private OnErrorListener onErrorListenerDelegate = new OnErrorListener() {
+    @Override public boolean onError(TrMediaPlayer mp, PlaybackException exception) {
+      return mOnErrorListener != null && mOnErrorListener.onError(mp, exception);
+    }
+  };
+
+  private OnInfoListener onInfoListenerDelegate = new OnInfoListener() {
+    @Override public boolean onInfo(TrMediaPlayer mp, PlaybackInfo info) {
+      return mOnInfoListener != null && mOnInfoListener.onInfo(mp, info);
+    }
+  };
+
+  private OnBufferingUpdateListener onBufferingUpdateListenerDelegate =
+      new OnBufferingUpdateListener() {
+        @Override public void onBufferingUpdate(TrMediaPlayer mp, int percent) {
+          if (mOnBufferingUpdateListener != null) {
+            mOnBufferingUpdateListener.onBufferingUpdate(mp, percent);
+          }
+        }
+      };
+
+  /**
+   * Resize the view based on the width and height specifications.
+   *
+   * @param widthMeasureSpec The specified width.
+   * @param heightMeasureSpec The specified height.
+   */
+  @Override protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+    super.onMeasure(widthMeasureSpec, heightMeasureSpec);
+    int width = getMeasuredWidth();
+    int height = getMeasuredHeight();
+    if (mVideoWidthHeightAspectRatio != 0) {
+      float viewAspectRatio = (float) width / height;
+      float aspectDeformation = mVideoWidthHeightAspectRatio / viewAspectRatio - 1;
+      if (aspectDeformation > MAX_ASPECT_RATIO_DEFORMATION_PERCENT) {
+        height = (int) (width / mVideoWidthHeightAspectRatio);
+      } else if (aspectDeformation < -MAX_ASPECT_RATIO_DEFORMATION_PERCENT) {
+        width = (int) (height * mVideoWidthHeightAspectRatio);
+      }
+    }
+    setMeasuredDimension(width, height);
+  }
+
+  /**
+   * Set the aspect ratio that this {@link VideoPlayerView} should satisfy.
+   *
+   * {@code Deprecated}, we gonna support ScaleType
+   *
+   * @param widthHeightRatio The width to height ratio.
+   */
+  @Deprecated public final void setVideoWidthHeightRatio(float widthHeightRatio) {
+    if (this.mVideoWidthHeightAspectRatio != widthHeightRatio) {
+      this.mVideoWidthHeightAspectRatio = widthHeightRatio;
+      requestLayout();
+    }
+  }
 
   public VideoPlayerView(Context context) {
     super(context);
-    initVideoView();
+    initialize(context);
   }
 
   public VideoPlayerView(Context context, AttributeSet attrs) {
     super(context, attrs);
-    initVideoView();
+    initialize(context);
   }
 
-  public VideoPlayerView(Context context, AttributeSet attrs, int defStyle) {
-    super(context, attrs, defStyle);
-    initVideoView();
+  public VideoPlayerView(Context context, AttributeSet attrs, int defStyleAttr) {
+    super(context, attrs, defStyleAttr);
+    initialize(context);
   }
 
   @TargetApi(Build.VERSION_CODES.LOLLIPOP)
   public VideoPlayerView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
     super(context, attrs, defStyleAttr, defStyleRes);
-    initVideoView();
+    initialize(context);
   }
 
-  @Override protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-    //Log.i("@@@@", "onMeasure(" + MeasureSpec.toString(widthMeasureSpec) + ", "
-    //        + MeasureSpec.toString(heightMeasureSpec) + ")");
-
-    int width = getDefaultSize(mVideoWidth, widthMeasureSpec);
-    int height = getDefaultSize(mVideoHeight, heightMeasureSpec);
-    if (mVideoWidth > 0 && mVideoHeight > 0) {
-
-      int widthSpecMode = MeasureSpec.getMode(widthMeasureSpec);
-      int widthSpecSize = MeasureSpec.getSize(widthMeasureSpec);
-      int heightSpecMode = MeasureSpec.getMode(heightMeasureSpec);
-      int heightSpecSize = MeasureSpec.getSize(heightMeasureSpec);
-
-      if (widthSpecMode == MeasureSpec.EXACTLY && heightSpecMode == MeasureSpec.EXACTLY) {
-        // the size is fixed
-        width = widthSpecSize;
-        height = heightSpecSize;
-
-        // for compatibility, we adjust size based on aspect ratio
-        if (mVideoWidth * height < width * mVideoHeight) {
-          //Log.i("@@@", "image too wide, correcting");
-          width = height * mVideoWidth / mVideoHeight;
-        } else if (mVideoWidth * height > width * mVideoHeight) {
-          //Log.i("@@@", "image too tall, correcting");
-          height = width * mVideoHeight / mVideoWidth;
-        }
-      } else if (widthSpecMode == MeasureSpec.EXACTLY) {
-        // only the width is fixed, adjust the height to match aspect ratio if possible
-        width = widthSpecSize;
-        height = width * mVideoHeight / mVideoWidth;
-        if (heightSpecMode == MeasureSpec.AT_MOST && height > heightSpecSize) {
-          // couldn't match aspect ratio within the constraints
-          height = heightSpecSize;
-        }
-      } else if (heightSpecMode == MeasureSpec.EXACTLY) {
-        // only the height is fixed, adjust the width to match aspect ratio if possible
-        height = heightSpecSize;
-        width = height * mVideoWidth / mVideoHeight;
-        if (widthSpecMode == MeasureSpec.AT_MOST && width > widthSpecSize) {
-          // couldn't match aspect ratio within the constraints
-          width = widthSpecSize;
-        }
-      } else {
-        // neither the width nor the height are fixed, try to use actual video size
-        width = mVideoWidth;
-        height = mVideoHeight;
-        if (heightSpecMode == MeasureSpec.AT_MOST && height > heightSpecSize) {
-          // too tall, decrease both width and height
-          height = heightSpecSize;
-          width = height * mVideoWidth / mVideoHeight;
-        }
-        if (widthSpecMode == MeasureSpec.AT_MOST && width > widthSpecSize) {
-          // too wide, decrease both width and height
-          width = widthSpecSize;
-          height = width * mVideoHeight / mVideoWidth;
-        }
-      }
-    } else {
-      // no size yet, just adopt the given spec sizes
-    }
-    setMeasuredDimension(width, height);
-  }
-
-  @Override public void onInitializeAccessibilityEvent(AccessibilityEvent event) {
-    super.onInitializeAccessibilityEvent(event);
-    event.setClassName(VideoPlayerView.class.getName());
-  }
-
-  @Override public void onInitializeAccessibilityNodeInfo(AccessibilityNodeInfo info) {
-    super.onInitializeAccessibilityNodeInfo(info);
-    info.setClassName(VideoPlayerView.class.getName());
-  }
-
-  public int resolveAdjustedSize(int desiredSize, int measureSpec) {
-    return getDefaultSize(desiredSize, measureSpec);
-  }
-
-  private void initVideoView() {
-    mVideoWidth = 0;
-    mVideoHeight = 0;
-    setSurfaceTextureListener(mSurfaceTextureListener);
+  private void initialize(Context context) {
+    // init
     setFocusable(true);
     setFocusableInTouchMode(true);
     requestFocus();
-    mCurrentState = STATE_IDLE;
-    mTargetState = STATE_IDLE;
+    this.mAudioCapabilities = AudioCapabilities.getCapabilities(context);
+    setSurfaceTextureListener(mSurfaceTextureListener);
   }
 
-  /**
-   * Sets video path.
-   *
-   * @param path the path of the video.
-   */
-  public void setMediaPath(String path) {
-    setMediaUri(Uri.parse(path));
+  private void releasePlayer() {
+    if (mMediaPlayer != null) {
+      mPlayerPosition = mMediaPlayer.getCurrentPosition();
+      mMediaPlayer.removeListener(mEventLogger);
+      mMediaPlayer.removeListener(mPlayerListener);
+      mMediaPlayer.release();
+      mMediaPlayer = null;
+      mEventLogger.endSession();
+      mEventLogger = null;
+      mExoMediaPlayerHelper = null;
+    }
   }
 
-  /**
-   * Sets video URI.
-   *
-   * @param uri the URI of the video.
-   */
-  public void setMediaUri(Uri uri) {
-    setMediaUri(uri, null);
+  private void preparePlayer(boolean playWhenReady) {
+    if (mUri == null || mSurface == null) {
+      return;
+    }
+
+    if (mMediaPlayer == null) {
+      mMediaPlayer = (ExoMediaPlayer) TrMediaPlayer.Factory.createExoPlayer(
+          RendererBuilderFactory.createRendererBuilder(getContext(), mUri));
+      mMediaPlayer.addListener(mPlayerListener);
+      mMediaPlayer.setOnPreparedListener(onPreparedListenerDelegate);
+      mMediaPlayer.setOnCompletionListener(onCompletionListenerDelegate);
+      mMediaPlayer.setOnErrorListener(onErrorListenerDelegate);
+      mMediaPlayer.setOnInfoListener(onInfoListenerDelegate);
+      mMediaPlayer.setOnBufferingUpdateListener(onBufferingUpdateListenerDelegate);
+
+      mMediaPlayer.setCaptionListener(mExoMediaPlayerHelper);
+      mMediaPlayer.setMetadataListener(mExoMediaPlayerHelper);
+      /* mMediaPlayer.setScreenOnWhilePlaying(true); */ // this is handle by this View
+      mMediaPlayer.seekTo(mPlayerPosition);
+      mPlayerNeedsPrepare = true;
+      mEventLogger = new EventLogger();
+      mEventLogger.startSession();
+      if (BuildConfig.DEBUG) {
+        mMediaPlayer.addListener(mEventLogger);
+        mMediaPlayer.setInfoListener(mEventLogger);
+        mMediaPlayer.setInternalErrorListener(mEventLogger);
+      }
+    }
+
+    if (mPlayerNeedsPrepare) {
+      mMediaPlayer.prepare();
+      mPlayerNeedsPrepare = false;
+    }
+
+    mMediaPlayer.setSurface(mSurface);
+    mMediaPlayer.setPlayWhenReady(playWhenReady);
+  }
+
+  @Override protected void onAttachedToWindow() {
+    super.onAttachedToWindow();
+    mAudioCapabilitiesReceiver =
+        new AudioCapabilitiesReceiver(getContext(), audioCapabilitiesListener);
+    mAudioCapabilitiesReceiver.register();
+  }
+
+  @Override protected void onDetachedFromWindow() {
+    super.onDetachedFromWindow();
+    releasePlayer();
+    mPlayerPosition = 0;
+    mAudioCapabilitiesReceiver.unregister();
+    mAudioCapabilitiesReceiver = null;
+
+    mAudioCapabilities = null;
+  }
+
+  private void maybeNotifyAudioCapabilitiesChanged(AudioCapabilities audioCapabilities) {
+    if (this.mAudioCapabilities == audioCapabilities) {
+      return;
+    }
+
+    if (mMediaPlayer == null) {
+      return;
+    }
+
+    boolean backgrounded = mMediaPlayer.getBackgrounded();
+    boolean playWhenReady = mMediaPlayer.getPlayWhenReady();
+    releasePlayer();
+    preparePlayer(playWhenReady);
+    mMediaPlayer.setBackgrounded(backgrounded);
+  }
+
+  @Override public void pause() {
+    mPlayRequested = false;
+    if (mMediaPlayer != null) {
+      mMediaPlayer.pause();
+    }
+  }
+
+  @Override public void stop() {
+    mPlayRequested = false;
+    releasePlayer();
+    mPlayerPosition = 0;
+  }
+
+  @Override public void seekTo(long milliSec) {
+    if (mMediaPlayer != null) {
+      mMediaPlayer.seekTo(milliSec);
+    }
   }
 
   @Override public void setVolume(@FloatRange(from = 0.f, to = 1.f) float volume) {
@@ -265,514 +420,79 @@ public class VideoPlayerView extends TextureView implements TrMediaPlayer.IMedia
     }
   }
 
-  /**
-   * Sets video URI using specific headers.
-   *
-   * @param uri the URI of the video.
-   * @param headers the headers for the URI request.
-   * Note that the cross domain redirection is allowed by default, but that can be
-   * changed with key/value pairs through the headers parameter with
-   * "android-allow-cross-domain-redirect" as the key and "0" or "1" as the value
-   * to disallow or allow cross domain redirection.
-   */
-  public void setMediaUri(Uri uri, Map<String, String> headers) {
-    mUri = uri;
-    mHeaders = headers;
-    mSeekWhenPrepared = 0;
-    openVideo();
-    requestLayout();
-    invalidate();
+  @TargetApi(23) private boolean requiresPermission(Uri uri) {
+    return Util.SDK_INT >= 23
+        && Util.isLocalFileUri(uri)
+        && ActivityCompat.checkSelfPermission(getContext(),
+        Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED;
   }
 
-  public void stopPlayback() {
-    if (mMediaPlayer != null) {
-      mMediaPlayer.stop();
-      mMediaPlayer.release();
-      mMediaPlayer = null;
-      mCurrentState = STATE_IDLE;
-      mTargetState = STATE_IDLE;
-      AudioManager am = (AudioManager) getContext().getApplicationContext()
-          .getSystemService(Context.AUDIO_SERVICE);
-      am.abandonAudioFocus(null);
+  @Override public void setMediaSource(@NonNull MediaSource source) {
+    if (source.mediaUri == null) {
+      throw new IllegalArgumentException("MediaSource must not be null");
     }
-  }
 
-  private void openVideo() {
-    if (mUri == null || mSurface == null) {
-      // not ready for playback just yet, will try again later
+    if (requiresPermission(source.mediaUri)) {
+      throw new RuntimeException("Permission to read this URI is not granted. "
+          + "Consider to request READ_EXTERNAL_STORAGE permission.");
+    }
+
+    if (this.mUri == source.mediaUri) {
       return;
     }
-    // we shouldn't clear the target state, because somebody might have
-    // called start() previously
-    release(false);
 
-    if (!mBackgroundAudioEnabled) {
-      AudioManager am = (AudioManager) getContext().getApplicationContext()
-          .getSystemService(Context.AUDIO_SERVICE);
-      am.requestAudioFocus(null, AudioManager.STREAM_MUSIC, AudioManager.AUDIOFOCUS_GAIN);
-    }
-
-    try {
-      //ExoMediaPlayer.RendererBuilder builder =
-      //    RendererBuilderFactory.createRendererBuilder(getContext(), mUri);
-      //mMediaPlayer = TrMediaPlayer.Factory.createExoPlayer(builder);
-      mMediaPlayer = TrMediaPlayer.Factory.createNativePlayer();
-
-      if (mAudioSession != 0) {
-        mMediaPlayer.setAudioSessionId(mAudioSession);
-      } else {
-        mAudioSession = mMediaPlayer.getAudioSessionId();
-      }
-
-      mMediaPlayer.setOnPreparedListener(mPreparedListener);
-      mMediaPlayer.setOnVideoSizeChangedListener(mSizeChangedListener);
-      mMediaPlayer.setOnCompletionListener(mCompletionListener);
-      mMediaPlayer.setOnErrorListener(mErrorListener);
-      mMediaPlayer.setOnInfoListener(mInfoListener);
-      mMediaPlayer.setOnBufferingUpdateListener(mBufferingUpdateListener);
-
-      mCurrentBufferPercentage = 0;
-      mMediaPlayer.setDataSource(getContext().getApplicationContext(), mUri, mHeaders);
-      mMediaPlayer.setSurface(mSurface);
-
-      // NOTE ExoPlayer's already dealt with this by MediaCodecAudioTrackRenderer
-      mMediaPlayer.setAudioStreamType(AudioManager.STREAM_MUSIC);
-      mMediaPlayer.setScreenOnWhilePlaying(true);
-      mMediaPlayer.prepareAsync();
-
-      // we don't set the target state here either, but preserve the
-      // target state that was there before.
-      mCurrentState = STATE_PREPARING;
-      attachMediaController();
-    } catch (IOException ex) {
-      Log.w(TAG, "Unable to open content: " + mUri, ex);
-      mCurrentState = STATE_ERROR;
-      mTargetState = STATE_ERROR;
-      mErrorListener.onError(mMediaPlayer,
-          new PlaybackException(MediaPlayer.MEDIA_ERROR_UNKNOWN, 0));
-      return;
-    } catch (IllegalArgumentException ex) {
-      Log.w(TAG, "Unable to open content: " + mUri, ex);
-      mCurrentState = STATE_ERROR;
-      mTargetState = STATE_ERROR;
-      mErrorListener.onError(mMediaPlayer,
-          new PlaybackException(MediaPlayer.MEDIA_ERROR_UNKNOWN, 0));
-      return;
-    }
+    this.mPlayerPosition = 0;
+    this.mUri = source.mediaUri;
+    mPlayRequested = false;
+    releasePlayer();
   }
 
-  public void setMediaController(TrMediaPlayer.Controller controller) {
-    if (mController != null) {
-      mController.hide();
-    }
-    mController = controller;
-    attachMediaController();
+  @Override public void setMediaUri(Uri uri) {
+    setMediaSource(new MediaSource(uri));
   }
 
-  private void attachMediaController() {
-    if (mMediaPlayer != null && mController != null) {
-      mController.setMediaPlayer(this);
-      View anchorView = this.getParent() instanceof View ? (View) this.getParent() : this;
-      mController.setAnchorView(anchorView);
-      mController.setEnabled(isInPlaybackState());
-    }
-  }
-
-  OnVideoSizeChangedListener mSizeChangedListener = new OnVideoSizeChangedListener() {
-    public void onVideoSizeChanged(TrMediaPlayer mp, int width, int height) {
-      mVideoWidth = mp.getVideoWidth();
-      mVideoHeight = mp.getVideoHeight();
-      if (mVideoWidth != 0 && mVideoHeight != 0) {
-        getSurfaceTexture().setDefaultBufferSize(mVideoWidth, mVideoHeight);
-        requestLayout();
-      }
-    }
-  };
-
-  OnPreparedListener mPreparedListener = new OnPreparedListener() {
-    public void onPrepared(TrMediaPlayer mp) {
-      mCurrentState = STATE_PREPARED;
-
-      if (mOnPreparedListener != null) {
-        mOnPreparedListener.onPrepared(mMediaPlayer);
-      }
-      if (mController != null) {
-        mController.setEnabled(true);
-      }
-
-      mVideoWidth = mp.getVideoWidth();
-      mVideoHeight = mp.getVideoHeight();
-
-      long seekToPosition =
-          mSeekWhenPrepared;  // mSeekWhenPrepared may be changed after seekTo() call
-      if (seekToPosition != 0) {
-        seekTo(seekToPosition);
-      }
-      if (mVideoWidth != 0 && mVideoHeight != 0) {
-        //Log.i("@@@@", "video size: " + mVideoWidth +"/"+ mVideoHeight);
-        getSurfaceTexture().setDefaultBufferSize(mVideoWidth, mVideoHeight);
-        // We won't get a "surface changed" callback if the surface is already the right size, so
-        // start the video here instead of in the callback.
-        if (mTargetState == STATE_PLAYING) {
-          start();
-          if (mController != null) {
-            mController.show();
-          }
-        } else if (!isPlaying() && (seekToPosition != 0 || getCurrentPosition() > 0)) {
-          if (mController != null) {
-            // Show the media controls when we're paused into a video and make 'em stick.
-            mController.show(0);
-          }
-        }
-      } else {
-        // We don't know the video size yet, but should start anyway.
-        // The video size might be reported to us later.
-        if (mTargetState == STATE_PLAYING) {
-          start();
-        }
-      }
-    }
-  };
-
-  OnCompletionListener mCompletionListener = new OnCompletionListener() {
-    public void onCompletion(TrMediaPlayer mp) {
-      mCurrentState = STATE_PLAYBACK_COMPLETED;
-      mTargetState = STATE_PLAYBACK_COMPLETED;
-      if (mController != null) {
-        mController.hide();
-      }
-      if (mOnCompletionListener != null) {
-        mOnCompletionListener.onCompletion(mMediaPlayer);
-      }
-    }
-  };
-
-  OnInfoListener mInfoListener = new OnInfoListener() {
-    public boolean onInfo(TrMediaPlayer mp, PlaybackInfo info) {
-      if (mOnInfoListener != null) {
-        mOnInfoListener.onInfo(mp, info);
-      }
-      return true;
-    }
-  };
-
-  OnErrorListener mErrorListener = new OnErrorListener() {
-    public boolean onError(TrMediaPlayer mp, PlaybackException er) {
-      Log.d(TAG, "Error: " + er.toString());
-      mCurrentState = STATE_ERROR;
-      mTargetState = STATE_ERROR;
-      if (mController != null) {
-        mController.hide();
-      }
-
-            /* If an error handler has been supplied, use it and finish. */
-      if (mOnErrorListener != null) {
-        if (mOnErrorListener.onError(mMediaPlayer, er)) {
-          return true;
-        }
-      }
-
-            /* Otherwise, pop up an error dialog so the user knows that
-             * something bad has happened. Only try and pop up the dialog
-             * if we're attached to a window. When we're going away and no
-             * longer have a window, don't bother showing the user an error.
-             */
-      if (getWindowToken() != null) {
-        Resources r = getContext().getResources();
-        int messageId;
-
-        if (er.what == MediaPlayer.MEDIA_ERROR_NOT_VALID_FOR_PROGRESSIVE_PLAYBACK) {
-          messageId = android.R.string.VideoView_error_text_invalid_progressive_playback;
-        } else {
-          messageId = android.R.string.VideoView_error_text_unknown;
-        }
-
-        new AlertDialog.Builder(getContext()).setMessage(messageId)
-            .setPositiveButton(android.R.string.VideoView_error_button,
-                new DialogInterface.OnClickListener() {
-                  public void onClick(DialogInterface dialog, int whichButton) {
-                                        /* If we get here, there is no onError listener, so
-                                         * at least inform them that the video is over.
-                                         */
-                    if (mOnCompletionListener != null) {
-                      mOnCompletionListener.onCompletion(mMediaPlayer);
-                    }
-                  }
-                })
-            .setCancelable(false)
-            .show();
-      }
-      return true;
-    }
-  };
-
-  OnBufferingUpdateListener mBufferingUpdateListener = new OnBufferingUpdateListener() {
-    public void onBufferingUpdate(TrMediaPlayer mp, int percent) {
-      Log.d(TAG,
-          "onBufferingUpdate() called with: " + "mp = [" + mp + "], percent = [" + percent + "]");
-      mCurrentBufferPercentage = percent;
-    }
-  };
-
-  /**
-   * Register a callback to be invoked when the media file
-   * is loaded and ready to go.
-   *
-   * @param l The callback that will be run
-   */
-  public void setOnPreparedListener(OnPreparedListener l) {
-    mOnPreparedListener = l;
-  }
-
-  /**
-   * Register a callback to be invoked when the end of a media file
-   * has been reached during playback.
-   *
-   * @param l The callback that will be run
-   */
-  public void setOnCompletionListener(OnCompletionListener l) {
-    mOnCompletionListener = l;
-  }
-
-  /**
-   * Register a callback to be invoked when an error occurs
-   * during playback or setup.  If no listener is specified,
-   * or if the listener returned false, TextureVideoView will inform
-   * the user of any errors.
-   *
-   * @param l The callback that will be run
-   */
-  public void setOnErrorListener(OnErrorListener l) {
-    mOnErrorListener = l;
-  }
-
-  /**
-   * Register a callback to be invoked when an informational event
-   * occurs during playback or setup.
-   *
-   * @param l The callback that will be run
-   */
-  public void setOnInfoListener(OnInfoListener l) {
-    mOnInfoListener = l;
-  }
-
-  public void setOnReleasedListener(OnReleasedListener listener) {
-    this.mOnReleaseListener = listener;
-  }
-
-  TextureView.SurfaceTextureListener mSurfaceTextureListener = new SurfaceTextureListener() {
-    @Override public void onSurfaceTextureSizeChanged(final SurfaceTexture surface, final int width,
-        final int height) {
-      boolean isValidState = (mTargetState == STATE_PLAYING);
-      boolean hasValidSize = (width > 0 && height > 0);
-      if (mMediaPlayer != null && isValidState && hasValidSize) {
-        if (mSeekWhenPrepared != 0) {
-          seekTo(mSeekWhenPrepared);
-        }
-        start();
-      }
-    }
-
-    @Override public void onSurfaceTextureAvailable(final SurfaceTexture surface, final int width,
-        final int height) {
-      mSurface = new Surface(surface);
-      openVideo();
-    }
-
-    @Override public boolean onSurfaceTextureDestroyed(final SurfaceTexture surface) {
-      // after we return from this we can't use the surface any more
-      if (mSurface != null) {
-        mSurface.release();
-        mSurface = null;
-      }
-      if (mController != null) mController.hide();
-      release(true);
-      return true;
-    }
-
-    @Override public void onSurfaceTextureUpdated(final SurfaceTexture surface) {
-      // do nothing
-    }
-  };
-
-  /*
-   * release the media player in any state
-   */
-  private void release(boolean clearTargetState) {
-    if (mMediaPlayer != null) {
-      if (this.mOnReleaseListener != null) {
-        mOnReleaseListener.onReleased(mUri, getCurrentPosition(), getDuration());
-      }
-      mMediaPlayer.reset();
-      mMediaPlayer.release();
-      mMediaPlayer = null;
-      mCurrentState = STATE_IDLE;
-      if (clearTargetState) {
-        mTargetState = STATE_IDLE;
-      }
-      AudioManager am = (AudioManager) getContext().getApplicationContext()
-          .getSystemService(Context.AUDIO_SERVICE);
-      am.abandonAudioFocus(null);
-    }
-  }
-
-  @Override public boolean onTouchEvent(MotionEvent ev) {
-    if (isInPlaybackState() && mController != null) {
-      toggleMediaControlsVisibility();
-    }
-    return false;
-  }
-
-  @Override public boolean onTrackballEvent(MotionEvent ev) {
-    if (isInPlaybackState() && mController != null) {
-      toggleMediaControlsVisibility();
-    }
-    return false;
-  }
-
-  @Override public boolean onKeyDown(int keyCode, KeyEvent event) {
-    boolean isKeyCodeSupported = keyCode != KeyEvent.KEYCODE_BACK &&
-        keyCode != KeyEvent.KEYCODE_VOLUME_UP &&
-        keyCode != KeyEvent.KEYCODE_VOLUME_DOWN &&
-        keyCode != KeyEvent.KEYCODE_VOLUME_MUTE &&
-        keyCode != KeyEvent.KEYCODE_MENU &&
-        keyCode != KeyEvent.KEYCODE_CALL &&
-        keyCode != KeyEvent.KEYCODE_ENDCALL;
-    if (isInPlaybackState() && isKeyCodeSupported && mController != null) {
-      if (keyCode == KeyEvent.KEYCODE_HEADSETHOOK || keyCode == KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE) {
-        if (mMediaPlayer.isPlaying()) {
-          pause();
-          mController.show();
-        } else {
-          start();
-          mController.hide();
-        }
-        return true;
-      } else if (keyCode == KeyEvent.KEYCODE_MEDIA_PLAY) {
-        if (!mMediaPlayer.isPlaying()) {
-          start();
-          mController.hide();
-        }
-        return true;
-      } else if (keyCode == KeyEvent.KEYCODE_MEDIA_STOP
-          || keyCode == KeyEvent.KEYCODE_MEDIA_PAUSE) {
-        if (mMediaPlayer.isPlaying()) {
-          pause();
-          mController.show();
-        }
-        return true;
-      } else {
-        toggleMediaControlsVisibility();
-      }
-    }
-
-    return super.onKeyDown(keyCode, event);
-  }
-
-  private void toggleMediaControlsVisibility() {
-    if (mController.isShowing()) {
-      mController.hide();
-    } else {
-      mController.show();
-    }
-  }
+  // IMediaPlayer
 
   @Override public void start() {
-    if (isInPlaybackState()) {
-      mMediaPlayer.start();
-      mCurrentState = STATE_PLAYING;
+    mPlayRequested = true;
+    if (mMediaPlayer == null) {
+      preparePlayer(true);
+    } else {
+      mMediaPlayer.setBackgrounded(false);
+      mMediaPlayer.setPlayWhenReady(mPlayRequested);
     }
-    mTargetState = STATE_PLAYING;
   }
 
   @Override public void start(long position) {
-    if (isInPlaybackState()) {
-      seekTo(position);
-      mMediaPlayer.start();
-      mCurrentState = STATE_PLAYING;
-    }
-    mTargetState = STATE_PLAYING;
-  }
-
-  @Override public void pause() {
-    if (isInPlaybackState()) {
-      if (mMediaPlayer.isPlaying()) {
-        mMediaPlayer.pause();
-        mCurrentState = STATE_PAUSED;
-      }
-    }
-    mTargetState = STATE_PAUSED;
-  }
-
-  @Override public void stop() {
-    // FIXME
-    pause();
-  }
-
-  public void suspend() {
-    release(false);
-  }
-
-  public void resume() {
-    openVideo();
+    seekTo(position);
+    start();
   }
 
   @Override public long getDuration() {
-    if (isInPlaybackState()) {
-      return mMediaPlayer.getDuration();
-    }
-
-    return -1;
+    return mMediaPlayer != null ? mMediaPlayer.getDuration() : ExoPlayer.UNKNOWN_TIME;
   }
 
   @Override public long getCurrentPosition() {
-    if (isInPlaybackState()) {
-      return mMediaPlayer.getCurrentPosition();
-    }
-    return 0;
-  }
-
-  @Override public void seekTo(long milliSec) {
-    if (isInPlaybackState()) {
-      mMediaPlayer.seekTo(milliSec);
-      mSeekWhenPrepared = 0;
-    } else {
-      mSeekWhenPrepared = milliSec;
-    }
+    return mMediaPlayer != null ? mMediaPlayer.getCurrentPosition() : ExoPlayer.UNKNOWN_TIME;
   }
 
   @Override public boolean isPlaying() {
-    return isInPlaybackState() && mMediaPlayer.isPlaying();
+    return mMediaPlayer != null && mMediaPlayer.isPlaying();
   }
 
   @Override public int getBufferPercentage() {
-    if (mMediaPlayer != null) {
-      return mCurrentBufferPercentage;
-    }
-    return 0;
+    return mMediaPlayer != null ? mMediaPlayer.getBufferedPercentage() : 0;
   }
 
-  private boolean isInPlaybackState() {
-    return (mMediaPlayer != null &&
-        mCurrentState != STATE_ERROR &&
-        mCurrentState != STATE_IDLE &&
-        mCurrentState != STATE_PREPARING);
-  }
-
-  public int getAudioSessionId() {
-    if (mAudioSession == 0) {
-      MediaPlayer player = new MediaPlayer();
-      mAudioSession = player.getAudioSessionId();
-      player.release();
-    }
-    return mAudioSession;
+  @Override public int getAudioSessionId() {
+    return mMediaPlayer != null ? mMediaPlayer.getAudioSessionId() : 0;
   }
 
   @Override public void setBackgroundAudioEnabled(boolean enabled) {
     mBackgroundAudioEnabled = enabled;
   }
 
-  @Override public void setMediaSource(@NonNull MediaSource source) {
-    setMediaUri(source.mediaUri);
+  private abstract class ExoMediaPlayerHelper
+      implements ExoMediaPlayer.CaptionListener, ExoMediaPlayer.Id3MetadataListener {
   }
 }
